# Improve performance and allow Uint8Array inputs