--- a/index.d.ts
+++ b/index.d.ts
@@ -5,6 +5,17 @@ export interface Options {
 	@default 32
 	*/
 	readonly size?: 32 | 64 | 128 | 256 | 512 | 1024;
+
+	/**
+	A Uint8Array used to encode the string into UTF-8 bytes.
+
+	This array can be reused across calls to `fnv1a`. Doing so will improve performance because it avoids allocating a new Uint8Array when encoding the string.
+
+	The size of the array does not have to be large enugh to hold the entire string, but performance will be improved if it is.
+
+	This option is only used when `value` is a string.
+	*/
+	readonly utf8Buffer?: Uint8Array;
 }
 
 /**
@@ -24,6 +35,14 @@ fnv1a('🦄🌈', {size: 128});
 
 Number(fnv1a('🦄🌈', {size: 32}));
 //=> 2868248295
+
+const utf8Buffer = new Uint8Array(100);
+fnv1a('🦄🌈', {size: 32, utf8Buffer});
+//=> 2868248295n
+
+const bytes = new Uint8Array([240, 159, 166, 132, 240, 159, 140, 136]);
+fnv1a(bytes, {size: 32});
+//=> 2868248295n
 ```
 */
-export default function fnv1a(string: string, options?: Options): bigint;
+export default function fnv1a(value: string | Uint8Array, options?: Options): bigint;