--- a/readme.md
+++ b/readme.md
@@ -25,15 +25,27 @@ fnv1a('🦄🌈', {size: 128});
 
 Number(fnv1a('🦄🌈', {size: 32}));
 //=> 2868248295
+
+const utf8Buffer = new Uint8Array(100);
+fnv1a('🦄🌈', {size: 32, utf8Buffer});
+//=> 2868248295n
+
+const bytes = new Uint8Array([240, 159, 166, 132, 240, 159, 140, 136]);
+fnv1a(bytes, {size: 32});
+//=> 2868248295n
 ```
 
 ## API
 
-### fnv1a(string, options?)
+### fnv1a(value, options?)
 
 Returns the hash as a positive `BigInt`.
 
 If you need it as a `number`, use `32` as `size` and wrap the return value in `Number(…)`.
+
+#### value
+
+Type: `string | Uint8Array`
 
 #### options
 
@@ -46,6 +58,18 @@ Values: `32 | 64 | 128 | 256 | 512 | 1024`\
 Default: `32`
 
 The bit size of the hash.
+
+##### utf8Buffer
+
+Type: `Uint8Array`
+
+A Uint8Array used to encode the string into UTF-8 bytes.
+
+This array can be reused across calls to `fnv1a`. Doing so will improve performance because it avoids allocating a new Uint8Array when encoding the string.
+
+The size of the array does not have to be large enugh to hold the entire string, but performance will be improved if it is.
+
+This option is only used when `value` is a string.
 
 ## Related