@sindresorhus 2023-11-16
Very nice! Thank you 🙏